Refactor option parsing. 1. Put all option parsing code in a separate file(options.cpp). 2. Store output file names or other intermediate parsing results in Globals. 3. Move Module::OutputType from Module to ispc.h.
