+++ Makefile
@@ -197,11 +197,13 @@ YACC=bison -d -v -t
 
 CORE_CXX_SRC=ast.cpp builtins.cpp cbackend.cpp ctx.cpp decl.cpp expr.cpp func.cpp \
 	ispc.cpp llvmutil.cpp module.cpp opt.cpp stmt.cpp sym.cpp \
-	type.cpp util.cpp
+	type.cpp util.cpp \
+    options.cpp
 ISPC_CXX_SRC=$(CORE_CXX_SRC) main.cpp
 CXX_SRC=$(ISPC_CXX_SRC)
 HEADERS=ast.h builtins.h ctx.h decl.h expr.h func.h ispc.h llvmutil.h module.h \
-	opt.h stmt.h sym.h type.h util.h
+	opt.h stmt.h sym.h type.h util.h \
+    options.h
 TARGETS=avx2-i64x4 avx11-i64x4 avx1-i64x4 avx1 avx1-x2 avx11 avx11-x2 avx2 avx2-x2 \
 	sse2 sse2-x2 sse4-8 sse4-16 sse4 sse4-x2 \
 	generic-4 generic-8 generic-16 generic-32 generic-64 generic-1 knl skx
